+++ language-support/hs/bindings/examples/nim-console/src/Logging.hs
@@ -1,7 +1,6 @@
 -- Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 -- SPDX-License-Identifier: Apache-2.0
 
-
 module Logging (Logger,noLog,tagLog,colourLog,plainLog,colourWrap) where
 
 import System.Console.ANSI(